refactor: split ui init and renderer
