+++ src/index.html
@@ -14,7 +14,7 @@
       </div>
     </div>
     <script type="module">
-      import { init } from '/game.ts';
+      import { init } from '/ui/init.ts';
       window.addEventListener('DOMContentLoaded', init);
     </script>
   </body>
